pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pantheon-server"
version = "0.1.0"
description = "FastAPI + Streamlit server for cryptocurrency analysis using pantheon-legends"
authors = [
    {name = "SpartanDigital", email = "spartandigital@icloud.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "pantheon-legends==0.2.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "streamlit>=1.28.0",
    "httpx>=0.25.0",
    "pandas>=2.1.0",
    "plotly>=5.17.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.11.0",
    "flake8>=6.1.0",
    "mypy>=1.7.0",
]

[project.urls]
"Homepage" = "https://github.com/SpartanDigitalDotNet/pantheon-server"
"Repository" = "https://github.com/SpartanDigitalDotNet/pantheon-server.git"
"Issues" = "https://github.com/SpartanDigitalDotNet/pantheon-server/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
